// BE/src/enums/crons.enum.ts
export enum CronStatus {
  PASS = 'PASS',
  FAIL = 'FAIL',
}

export enum CRONTYPES {
  BOOSTER_DEACTIVATION = 'Booster Deactivation',
  INCENTIVE_POLICY_DEACTIVATION = 'Incentive Policy Deactivation',
  BOOKING_USER_REFRESH = 'Booking And User Refresh',
  NOTIFICATIONS = 'Notifications',
  REPORTS = 'Reports',
  ACCRUALS = 'Accruals',
  RM_DASHBOARD_DAILY_REPORT = 'RM Dashboard Daily Report',
  OFFICE_USE_REMINDER = 'Office Use Reminder',
}